This is the game hope you like it
I have imported the random library and used basic functions such as while loops and if and else statements in this project. In the list, I have listed the names of some of my favourite anime's. You always modify the project and please do share the modified code with everyone. Have a nice day.